Unify the two managers' parsers
